+++ config/urls.py
@@ -20,5 +20,6 @@ from accounts.views import *
 
 urlpatterns = [
     path("admin/", admin.site.urls),
+    path('oauth/', include('oauths.urls')),
     path('account/', include('accounts.urls')),
 ]
